backend2/requirements.txt
tensorflow==2.10.0
scikit-learn==1.3.0
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.13.0
pandas==2.1.0
fastapi
uvicorn
pillow
python-multipart
httpx
exifread
gdown